diversity_sampler_with_viz/requirements.txt
numpy>=1.21
scikit-learn>=1.2
# optional for UMAP
umap-learn>=0.5
matplotlib>=3.6
